services/anomaly/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
tensorflow
joblib
python-multipart
kafka-python
